refactor(architecture): one folder for containers
