+++ src/components/containers/ContributorsContainer.jsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 import LoadingWrapper from './../atoms/LoadingWrapper'
 import { getDataList } from './../../services/GithubApiService'
 import { setCache, getCache } from './../../services/CacheService'
-import Contributors from './../organisms/Contributors'
+import Contributors from './../ecosystems/Contributors'
 import Paginator from './../atoms/Paginator'
 import { CACHE_TYPE_CONTRIBUTORS, CACHE_DURATION_MINUTE } from './../../constants/RepositoryConstants'
 import Error from './../atoms/Error'
