+++ src/components/ecosystems/styles/Contributors.sass
